+++ .eslintrc.js
@@ -18,7 +18,10 @@ module.exports = {
   rules: {
     "comma-dangle": [2, "only-multiline"],
     "no-console": 0,
-    "no-extra-boolean-cast": 0
+    "no-extra-boolean-cast": 0,
+    // warn against future uses of jQuery
+    // TODO: set this to level 2 (error) once existing uses are replaced
+    "ember/no-jquery": 1
   },
   globals: {
     "_": false,
@@ -62,6 +65,19 @@ module.exports = {
       excludedFiles: ['tests/dummy/**/*.js'],
       env: {
         embertest: true
+      },
+      rules: {
+        // TODO: bump this to level 1 (warn) once we're ready to remove jQuery from tests
+        "ember/no-jquery": 0
+      }
+    },
+
+    // dummy app files
+    {
+      files: ['tests/dummy/**/*.js'],
+      rules: {
+        // TODO: bump this to level 1 (warn) once we're ready to remove jQuery from the dummy app
+        "ember/no-jquery": 0
       }
     }
   ]